Created: - constantStat.html; otherStat.html; serviceStat.html - ConstantCostService; OtherExpensesService; ServiceEntriesService - ConstantCostServiceImpl; OtherExpensesServiceImplServiceIm; ServiceEntriesServiceImpl
+ little corrections

+++ src/main/java/pl/edu/wszib/ce/carexpenses/service/FuelEntriesServiceImpl.java
@@ -55,7 +55,7 @@ public class FuelEntriesServiceImpl implements FuelEntriesService {
 
     @Override
     public Float avgFuelDist() {
-        return fuelEntriesDao.avgFuelDist().floatValue();
+        return fuelEntriesDao.avgFuelDist();
     }
 
     @Override
